feat: add delete item from cart
Change the UI of adding and reducing the quantity of a product from cart
by removing the use of input and just using a group of buttons.

Add feature of removing items from cart by filtering the global products
state from zustand and reducing the total cost and items added to cart.